requirements.txt
# ML and RL framework
stable-baselines3>=2.0.0
gymnasium>=0.28.0
numpy>=1.21.0
pandas>=1.3.0

# Google Cloud
google-cloud-aiplatform>=1.34.0
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.0
google-cloud-pubsub>=2.18.0

# Web framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Cryptocurrency data
ccxt>=4.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
requests>=2.31.0

# Logging and monitoring
google-cloud-logging>=3.8.0